requirements.txt
pandas
playwright
unidecode
pyyaml
